Fix import shenanigans 2/2 (#3497)
* refactor regs module

* add docs about common import pitfalls

* make pwndbg.aglib.proc a normal submodule

also make pwndbg.lib.memory not depend on
aglib

* prevent import cycle

* add comment about importing commands

* finish pwndbg.lib.memory fixup

* dont use aglib in lib

* clearly i need to clarify

* fixups all around

* a good life

* cleanup docs

* remove unused imports

* fix nextproginstr

* Apply suggestion from @disconnect3d

* upd with more pitfalls

* fixup from pwndbg.aglib import regs doc

its not really a copy of the object, in this case `regs`
would be bound to the same object as `aglib.regs`
(None) and changes in one would reflect the other
if it wasn't None.

It is a kind of "copy of the name", which is what
you were going for, but maybe this edit makes it
a bit less ambiguous.

* fix read_reg(cs)

* pc -> eip. use the pc property if you want the pc!

* pitfalls doc imports

* fix tests and fix tests

* flew too close to the sun

* okay I see the issue

* remove duplicated gdb test

* fix lldb by setting proper event priorities

so satisfying

* remove unneeded gdb test

again

* move this troublesome line to PR 3/2

---------

Co-authored-by: Disconnect3d <dominik.b.czarnota@gmail.com>