Fill full height of context section in TUI mode (#3511)
* Fill full height of context section in TUI mode

Add new `context-tui-adjust-height` config option to dynamically increase
the number of lines printed in `regs`, `disasm`, `code`, and `threads` sections
to fill the whole TUI window allocated for the section.

This ignores the other config values in place to limit the number of lines
for each individual section and interacts weirdly with the context history,
because the same entry could be added twice with longer section output
when switching from CLI to TUI mode and the context was already output
in CLI mode. That is a future TODO to replace the section data in the latest
history entry when applicable.

* Improve help strings and regenerate docs

* Add typehints to context commands

* Add type hints for context output redirection

The CallOutput redirection doesn't fully implement the TextIO protocol but only the necessary functions used in the code. Instead of implementing all of them keep a type union around.